+++ x-pack/plugins/security_solution/public/detection_engine/rule_management/logic/bulk_actions/use_all_rule_ids_for_bulk_action.ts
